backend/requirements.txt
langchain
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
python-dotenv
fastapi-cors